+++ backend/simcheong2/src/main/java/com/example/simcheong2/global/redis/repository/RedisTokensRepository.java
@@ -1,6 +1,6 @@
-package com.example.simcheong2.global.security.redis.repository;
+package com.example.simcheong2.global.redis.repository;
 
-import com.example.simcheong2.global.security.redis.entity.RedisTokens;
+import com.example.simcheong2.global.redis.entity.RedisTokens;
 import org.springframework.data.repository.CrudRepository;
 
 public interface RedisTokensRepository extends CrudRepository<RedisTokens, String> {
